% Generated by roxygen2 (4.0.1): do not edit by hand
\name{photoz.diagnostic}
\alias{photoz.diagnostic}
\title{Diagnostic for  photometric redshift fit}
\usage{
photoz.diagnostic(photoz, specz)
}
\arguments{
\item{photoz}{data.frame}

\item{specz}{data.frame}
}
\value{
data.frame
}
\description{
Diagnostic for  photometric redshift fit
}
\examples{
y <- rgamma(100,10,.1)
summary(glm(y~1,family=Gamma))
}

