\name{CosmoPhotoz-package}
\alias{CosmoPhotoz-package}
\alias{CosmoPhotoz}
\docType{package}
\title{
Photometric redshift estimation based on  Generalized Linear Models
}
\description{
This package provides a user-friendly interface to perform fast and reliable  photometric redshift estimation. The code make use generalized linear models from a Gamma family, both from a frequentist and a bayesian perspective. 
}
\details{
\tabular{ll}{
Package: \tab CosmoPhotoz\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-15\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Rafael S. de Souza

Maintainer: Rafael S. de Souza <rafael.2706@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
