% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CosmoPhotoz}
\alias{CosmoPhotoz}
\title{Photo-z estimation from a training dataset and a test dataset}
\usage{
CosmoPhotoz(trainData, testData, numberOfPcs=4, method="Bayesian", family="gamma")
}
\arguments{
\item{trainData}{vector containing spectroscopic redshift data and photometry}

\item{testData}{vector containing spectroscopic redshift data and photometry}

\item{numberOfPcs}{an integer indicating the number of principal components to consider}

\item{method}{a string containing the chosen GLM method. Two options are available: \code{Frequentist} will use the function  \code{\link{glm}} from the package \code{stats}; \code{Bayesian} will use the function \code{\link{bayesglm}} from the package  \code{arm}.}

\item{family}{a string containing \code{gamma} or \code{inverse.gaussian} (a description of the error distribution and link function to be used in the model).}
}
\value{
a vector with the estimated photometric redshifts
}
\description{
\code{CosmoPhotoz} returns photometric redshift estimated from photometric
data and a training dataset with photometry and spectroscopy. The estimation is
based on GLMs (see \link{\code{glmTrainPhotoz}} and \link{\code{glmPredictPhotoz}}).
}
\examples{
# First, generate some mock data
}
\author{
Alberto Krone-Martins, Rafael S. de Souza
}
\keyword{utilities}

