% Generated by roxygen2 (4.0.1): do not edit by hand
\name{glmPredictPhotoZ}
\alias{glmPredictPhotoZ}
\title{Predict photometric redshift from a GLM fit}
\usage{
glmPredictPhotoZ(data, train)
}
\arguments{
\item{x}{data.frame}

\item{train}{GLM object}
}
\value{
list
}
\description{
Predict photometric redshift from a GLM fit
}
\examples{
y <- rgamma(100,10,.1)
summary(glm(y~1,family=Gamma))
}

