% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{PHAT0train}
\alias{PHAT0train}
\title{PHAT0 train dataset}
\format{A dataframe with 8478 observations on the following on 12 variables, no
  NAs.

\describe{

\item{list("redshift")}{The redshift of the galaxy.}
\item{list("up")}{The magnitude of the galaxy in the up band.}
\item{list("gp")}{The magnitude of the galaxy in the gp band.}
\item{list("rp")}{The magnitude of the galaxy in the rp band.}
\item{list("ip")}{The magnitude of the galaxy in the ip band.}
\item{list("zp")}{The magnitude of the galaxy in the zp band.}
\item{list("Y")}{The magnitude of the galaxy in the Y band.}
\item{list("J")}{The magnitude of the galaxy in the J band.}
\item{list("H")}{The magnitude of the galaxy in the H band.}
\item{list("K")}{The magnitude of the galaxy in the K band.}
\item{list("IRAC_1")}{The magnitude of the galaxy in the IRAC_1 band.}
\item{list("IRAC_2")}{The magnitude of the galaxy in the IRAC_2 band.}

}}
\source{
Hildebrandt H., Arnouts S., Capak P., Moustakas L. A., Wolf C., Abdalla
  F. B., Assef R. J., Banerji M., et. al (2010), Astronomy&Astrophysics, v.523, p.A31
}
\description{
The CosmoPhotoz package includes the dataset \code{PHAT0train}, containing photometric
data in the astronomical photometric bands up, gp, rp, ip, zp, Y, J, H, K, IRAC_1, IRAC_2
and the redshift, for 8478 galaxies. The original PHAT0 dataset, containing 169520 was
split in two (\code{\link{PHAT0train}} and \code{\link{PHAT0test}}).
}
\keyword{datasets}

