% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_photoz}
\alias{plot_photoz}
\title{Plot Predict photometric vs observed  redshift from a GLM fit}
\usage{
plot_photoz(photoz, specz, type = c("errordist", "predobvs"))
}
\arguments{
\item{photoz}{data.frame}

\item{specz}{data.frame}
}
\value{
ggplot object
}
\description{
Plot Predict photometric vs observed  redshift from a GLM fit
}
\examples{
y <- rgamma(100,10,.1)
summary(glm(y~1,family=Gamma))
}

