% Generated by roxygen2 (4.0.1): do not edit by hand
\name{TrainGLM}
\alias{TrainGLM}
\title{Fit a GLM function in the training set}
\usage{
TrainGLM(x, method = c("Frequentist", "Bayesian"), family = c("gamma",
  "inverse.gaussian"))
}
\arguments{
\item{x}{data.frame}
}
\value{
GLM object
}
\description{
Fit a GLM function in the training set
}
\examples{
y <- rgamma(100,10,.1)
summary(glm(y~1,family=Gamma))
}

