% Generated by roxygen2 (4.0.1): do not edit by hand
\name{computeDiagPhotoZ}
\alias{computeDiagPhotoZ}
\title{Diagnostic for  photometric redshift fit}
\usage{
computeDiagPhotoZ(photoz, specz)
}
\arguments{
\item{photoz}{vector}

\item{specz}{vector}
}
\value{
list
}
\description{
Diagnostic for  photometric redshift fit
}
\examples{
y <- rgamma(100,10,.1)
summary(glm(y~1,family=Gamma))
}

