% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Predict.redshift}
\alias{Predict.redshift}
\title{Predict photometric redshift from a GLM fit}
\usage{
Predict.redshift(data = x, train = train)
}
\arguments{
\item{x}{data.frame}

\item{train}{GLM object}
}
\value{
data.frame
}
\description{
Predict photometric redshift from a GLM fit
}
\examples{
y <- rgamma(100,10,.1)
summary(glm(y~1,family=Gamma))
}

