% Generated by roxygen2 (4.0.1): do not edit by hand
\name{glmPredictPhotoZ}
\alias{glmPredictPhotoZ}
\title{Predict photometric redshifts using a given a GLM fit object}
\usage{
glmPredictPhotoZ(data, train)
}
\arguments{
\item{data}{a data.frame containing the data one wished to compute the redshift}

\item{train}{a trainned GLM object containing the fit of the model}
}
\value{
list containing the results of the redshift estimation
}
\description{
\code{glmPredictPhotoZ} computes a list of simple summary
statistics for the photometric redshift estimation.
}
\examples{
\dontrun{
# First, generate some mock data
}
}
\author{
Rafael S. de Souza, Alberto Krone-Martins
}
\keyword{utilities}

