% Generated by roxygen2 (4.0.1): do not edit by hand
\name{glmTrainPhotoZ}
\alias{glmTrainPhotoZ}
\title{Fit a GLM function in the training set}
\usage{
glmTrainPhotoZ(x, formula = NULL, method = c("Frequentist", "Bayesian"),
  family = c("quantile", "gamma", "inverse.gaussian"))
}
\arguments{
\item{x}{data.frame}
}
\value{
GLM object
}
\description{
Fit a GLM function in the training set
}
\examples{
y <- rgamma(100,10,.1)
summary(glm(y~1,family=Gamma))
}

