% Generated by roxygen2 (4.0.1): do not edit by hand
\name{glmTrainPhotoZ}
\alias{glmTrainPhotoZ}
\title{Fit a GLM for photometric redshift estimation}
\usage{
glmTrainPhotoZ(x, formula=NULL, method=c("Frequentist","Bayesian"), family=c("gamma","inverse.gaussian"))
}
\arguments{
\item{x}{a data.frame containing the data to train the model}

\item{formula}{an object of class "formula" to be adopted}

\item{method}{a string containing the chosen GLM method. Two options are available: \code{Frequentist} will use the function  \code{\link{glm}} from the package \code{\link{mixlm}}; \code{Bayesian} will use the function \code{\link{bayesglm}} from the package  \code{\link{arm}}.}

\item{family}{a string containing \code{gamma} or \code{inverse.gaussian} (a description of the error distribution and link function to be used in the model).}
}
\value{
GLM object a trainned GLM object containing the fit of the model
}
\description{
\code{glmTrainPhotoZ} trains a generalized linear model for
photometric redshift estimation.
}
\examples{
\dontrun{
# First, generate some mock data
ppo <- runif(1000, min=0.1, max=2)
ppo_ph <- rnorm(length(ppo), mean=ppo, sd=0.05)

# Now, mock a redshift training and estimation
}
}
\author{
Rafael S. de Souza, Alberto Krone-Martins
}
\keyword{utilities}

