% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotDiagPhotoZ}
\alias{plotDiagPhotoZ}
\title{Plot Predict photometric vs observed redshift from a GLM fit

\code{plotDiagPhotoZ} returns diagnostic plots from the results of
photometric redshifts. Different types of plots are available. The
produced plots will be returned as ggplot objects.}
\usage{
plotDiagPhotoZ(photoz, specz, type = c("errordist", "predobs", "errorviolins",
  "box"))
}
\arguments{
\item{photoz}{vector}

\item{specz}{vector}

\item{type}{list}
}
\value{
ggplot object
}
\description{
Plot Predict photometric vs observed redshift from a GLM fit

\code{plotDiagPhotoZ} returns diagnostic plots from the results of
photometric redshifts. Different types of plots are available. The
produced plots will be returned as ggplot objects.
}
\examples{
\dontrun{
sum("a")
}
}

